Made a lot of changes added HOC

+++ src/services/apiCabins.js
@@ -1,46 +1,68 @@
 import supabase, { supabaseUrl } from "./supabase";
 
 export async function getCabins() {
-  let { data, error } = await supabase.from("cabins").select("*");
+  const { data, error } = await supabase.from("cabins").select("*");
 
   if (error) {
     console.error(error);
-    throw new Error("Cabins data cannot be loaded..");
+    throw new Error("Cabins could not be loaded");
   }
+
   return data;
 }
 
-export async function createCabin(newCabin) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/Cabins/${imageName}`;
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/Cabins/${imageName}`;
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
 
-  //Creating cabin
-  const { data, error } = await supabase.from("cabins").insert([newCabin]);
-  //upload image
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
+
+  if (error) {
+    console.error(error);
+    throw new Error("Cabin could not be created");
+  }
+
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("Cabins")
     .upload(imageName, newCabin.image);
 
-  //Delete cabin if there was an error
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
-    throw new error(
-      "Cabin image coulld not be uploaded and cabin could not be created"
+    throw new Error(
+      "Cabin image could not be uploaded and the cabin was not created"
     );
   }
+
   return data;
 }
 
 export async function deleteCabin(id) {
   const { data, error } = await supabase.from("cabins").delete().eq("id", id);
+
   if (error) {
     console.error(error);
-    throw new error("Cabins cannot be deleted");
+    throw new Error("Cabin could not be deleted");
   }
 
   return data;
